pyproject.toml
[project]
name = "opendialect"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "asyncio>=4.0.0",
    "gradio>=6.8.0",
    "openai>=2.24.0",
    "openai-agents>=0.10.3",
    "python-dotenv>=1.2.2",
]
